+++ lib/core/router.dart
@@ -4,6 +4,8 @@ import 'package:mongbi_app/core/custom_transition_page.dart';
 import 'package:mongbi_app/core/main_scaffold.dart';
 import 'package:mongbi_app/core/responsive_layout.dart';
 import 'package:mongbi_app/core/route_observer.dart';
+import 'package:mongbi_app/features/alarm/presentation/pages/remind_time_setting_page.dart';
+import 'package:mongbi_app/features/alarm/presentation/pages/remind_time_time_setting_page.dart';
 import 'package:mongbi_app/features/auth/presentation/pages/social_login_page.dart';
 import 'package:mongbi_app/features/challenge/presentation/pages/challenge_intro_page.dart';
 import 'package:mongbi_app/features/challenge/presentation/pages/challenge_page.dart';
@@ -17,8 +19,6 @@ import 'package:mongbi_app/presentation/alarm/alarm_page.dart';
 import 'package:mongbi_app/presentation/home/home_page.dart';
 import 'package:mongbi_app/presentation/onboarding/onbording_exit_page.dart';
 import 'package:mongbi_app/presentation/onboarding/onbording_page.dart';
-import 'package:mongbi_app/presentation/remind/remind_time_setting_page.dart';
-import 'package:mongbi_app/presentation/remind/remind_time_time_setting_page.dart';
 import 'package:mongbi_app/presentation/setting/alarm_setting_page.dart';
 import 'package:mongbi_app/presentation/setting/nickname_input_page.dart';
 import 'package:mongbi_app/presentation/setting/profile_setting_page.dart';
