+++ lib/features/alarm/data/data_sources/alarm_data_source.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/data/dtos/alarm_dto.dart';
+import 'package:mongbi_app/features/alarm/data/dtos/alarm_dto.dart';
 
 abstract interface class AlarmDataSource {
   Future<List<AlarmDto>?> fetchAlarms();
