+++ lib/features/alarm/domain/use_cases/update_is_read_status_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/domain/repositories/alarm_repository.dart';
+import 'package:mongbi_app/features/alarm/domain/repositories/alarm_repository.dart';
 
 class UpdateIsReadStatusUseCase {
   UpdateIsReadStatusUseCase(this.repository);
