+++ lib/features/alarm/presentation/pages/remind_time_time_setting_page.dart
@@ -5,12 +5,12 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/analytics/analytics_helper.dart';
 import 'package:mongbi_app/core/font.dart';
+import 'package:mongbi_app/features/alarm/presentation/pages/widgets/remind_time_setting_text_widget.dart';
+import 'package:mongbi_app/features/alarm/presentation/pages/widgets/remind_time_setting_widget.dart';
+import 'package:mongbi_app/features/alarm/presentation/view_models/remind_time_setting_view_model.dart';
 import 'package:mongbi_app/presentation/common/button_type.dart';
 import 'package:mongbi_app/presentation/common/custom_snack_bar.dart';
 import 'package:mongbi_app/presentation/common/filled_button_widget.dart';
-import 'package:mongbi_app/presentation/remind/view_model/remind_time_setting_view_model.dart';
-import 'package:mongbi_app/presentation/remind/widgets/remind_time_setting_text_widget.dart';
-import 'package:mongbi_app/presentation/remind/widgets/remind_time_setting_widget.dart';
 import 'package:mongbi_app/providers/setting_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
