+++ lib/presentation/alarm/widgets/delete_modal.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/font.dart';
+import 'package:mongbi_app/features/alarm/providers/alarm_provider.dart';
 import 'package:mongbi_app/presentation/common/button_type.dart';
 import 'package:mongbi_app/presentation/common/filled_button_widget.dart';
 import 'package:mongbi_app/presentation/common/ghost_button_widget.dart';
-import 'package:mongbi_app/providers/alarm_provider.dart';
 
 class DeleteModal extends StatelessWidget {
   const DeleteModal({super.key});
