+++ lib/presentation/setting/alarm_setting_page.dart
@@ -8,7 +8,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/analytics/analytics_helper.dart';
 import 'package:mongbi_app/core/font.dart';
-import 'package:mongbi_app/presentation/remind/view_model/remind_time_setting_view_model.dart';
+import 'package:mongbi_app/features/alarm/presentation/view_models/remind_time_setting_view_model.dart';
 import 'package:mongbi_app/presentation/setting/widgets/setting_rounded_list_tile_item.dart';
 import 'package:mongbi_app/presentation/setting/widgets/setting_toggle_switch.dart';
 import 'package:mongbi_app/providers/setting_provider.dart';
